Add after_downloads action to gpush_core
For now this is a simple method the client needs to call after
it has done the downloads it wants. This is NOT a great API but
it gets us where we want with lint.py. In future we'll have
a declarative approach to driving downloads and options like making
executable, and for lifecycle code that will run more automatically.